# openapi.yaml
openapi: 3.1.0
info:
  title: Knowledge Management System API
  version: 1.0.0
  description: >
    Upload text and image files, have Claude generate searchable metadata
    (description, tags, keywords) on upload, and search assets by keyword over
    that metadata. All endpoints are served under `/api`.
servers:
  - url: /api
    description: Same-origin API (the server also serves the built web client)

paths:
  /health:
    get:
      summary: Health check
      description: Reports service health and whether an AI (Anthropic) key is configured.
      operationId: getHealth
      responses:
        "200":
          description: Service is up.
          content:
            application/json:
              schema:
                type: object
                required: [status, ai]
                properties:
                  status:
                    type: string
                    example: ok
                  ai:
                    type: boolean
                    description: True when ANTHROPIC_API_KEY is configured (metadata generation active).
                    example: true

  /assets:
    get:
      summary: List all assets
      description: Returns every asset, newest first.
      operationId: listAssets
      responses:
        "200":
          description: All assets.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      summary: Upload a file
      description: >
        Upload a single text or image file (multipart field `file`). The file is
        stored and Claude generates metadata on upload. If AI is unavailable or no
        key is set, the asset is still created with empty metadata and
        `aiGenerated: false`.
      operationId: uploadAsset
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: >
                    A text file (`text/*`) or an image in a Claude-vision-supported
                    format (`image/jpeg`, `image/png`, `image/gif`, `image/webp`).
                    Max 15 MB. Other types return 400.
      responses:
        "201":
          description: Asset created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "400":
          description: No file, unsupported type, or file too large.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /assets/search:
    get:
      summary: Keyword search
      description: >
        Ranked keyword search over each asset's description, tags, keywords,
        extracted text, and original filename. AND semantics: every whitespace-
        separated token in `q` must appear (case-insensitive). Results are ranked
        by total term occurrences. An empty query returns an empty array.
      operationId: searchAssets
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Query string, e.g. `black hair` or `document`.
          example: black hair
      responses:
        "200":
          description: Matching assets, ranked most-relevant first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"

  /assets/{id}:
    get:
      summary: Get one asset
      operationId: getAsset
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: The asset.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: No asset with that id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete an asset
      description: Removes the asset's metadata row and its stored file.
      operationId: deleteAsset
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "204":
          description: Deleted.
        "404":
          description: No asset with that id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /assets/{id}/reevaluate:
    post:
      summary: Re-generate AI metadata for an asset
      description: >
        Re-runs Claude on the stored file and overwrites the asset's description,
        tags, and keywords. Useful for assets whose metadata generation failed
        transiently (e.g. before credits were available).
      operationId: reevaluateAsset
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: The updated asset.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "400":
          description: >
            The asset is an image in a format Claude vision cannot process
            (only jpeg/png/gif/webp are supported).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Asset or underlying file not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /assets/{id}/raw:
    get:
      summary: Download the raw file
      description: Serves the stored file bytes with its original content type (image bytes or text).
      operationId: getAssetRaw
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: The file contents.
          content:
            image/*:
              schema:
                type: string
                format: binary
            text/*:
              schema:
                type: string
        "404":
          description: Asset or underlying file not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  parameters:
    AssetId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Asset id (nanoid).
      example: 0VujuIu36UHmc_h6lWuEN

  schemas:
    Asset:
      type: object
      required:
        - id
        - kind
        - originalName
        - storedName
        - mimeType
        - sizeBytes
        - createdAt
        - description
        - tags
        - keywords
        - extractedText
        - aiGenerated
      properties:
        id:
          type: string
          description: Unique id (nanoid).
        kind:
          type: string
          enum: [image, text]
        originalName:
          type: string
          description: Filename the user uploaded.
        storedName:
          type: string
          description: Filename on disk under the uploads directory.
        mimeType:
          type: string
        sizeBytes:
          type: integer
        createdAt:
          type: string
          format: date-time
        description:
          type: string
          description: AI-generated description (empty when aiGenerated is false).
        tags:
          type: array
          items:
            type: string
          description: AI-generated broad-category tags.
        keywords:
          type: array
          items:
            type: string
          description: AI-generated specific searchable terms.
        extractedText:
          type: string
          description: Full text for text files; empty for images. Also searchable.
        aiGenerated:
          type: boolean
          description: True when Claude produced the metadata; false if AI was unavailable.
      example:
        id: 0VujuIu36UHmc_h6lWuEN
        kind: text
        originalName: doc-note.txt
        storedName: 0VujuIu36UHmc_h6lWuEN.txt
        mimeType: text/plain
        sizeBytes: 107
        createdAt: "2026-07-17T10:57:09.675Z"
        description: Meeting notes requesting review of a quarterly financial document and a signed contract.
        tags: [meeting notes, finance, contracts]
        keywords: [quarterly financial document, contract signing, deadline]
        extractedText: "Meeting notes: please review the attached quarterly financial document and sign the contract before Friday."
        aiGenerated: true

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
      example:
        error: "unsupported file type: application/pdf"